Merged PR 2697: Education Details Page
The education details page can be accessed with http://localhost:3000/education/1 and http://localhost:3000/education/2.
The numbers 1 and 2 are key/id set for accessing a particular blog/video. They are set in the file **DetailsContent.ts**.
![image.png](https://mintyfusion.visualstudio.com/67c7b9c7-621a-4e57-bfc0-201777dfde3d/_apis/git/repositories/f8dee45e-f292-4037-9538-f2b59bef82ef/pullRequests/2697/attachments/image.png)
As of now the breadcrumb, category, title and image are set. The rest of the contents of the blog/video will be coming from Agility in HTML.